Refatora código do maincontent, home. Adiciona integração de upload de perfil de startup, e intergação com endpoint de listagem de sócios
