+++ src/app/components/ContentStartup/contentstartup.jsx
@@ -15,6 +15,7 @@ import updateDetailsStartup from "../../api/put/update-details-startup";
 import CardDetails from "../CardDetails/CardDetails";
 
 import { ProfileCArdImageStartup } from "../ProfileCardImageStartup/ProfileCardImageStartup";
+import profileImageStartup from "../../api/post/profile-image-startup";
 
 
 export function ContentStartup() {
@@ -44,15 +45,16 @@ export function ContentStartup() {
     }, [enterprise_id, token, reset]);
 
     const handleProfileImageChange = async (e) => {
+        console.log("ativada")
         const file = e?.target?.files[0];
         if (file) {
             const formData = new FormData();
             formData.append("profile_picture", file);
-
-            const response = await addProfileImage(user?.id, token, formData);
-
+    
+            const response = await profileImageStartup(enterprise_id, formData, token);
+    
             if (response) {
-                console.log('Imagem de perfil atualizada com sucesso', response);
+            console.log('Imagem de perfil atualizada com sucesso', response);
             }
         }
         window.location.reload();
@@ -75,11 +77,10 @@ export function ContentStartup() {
         updateDetailsStartup(enterprise_id, data, token)
             .then((response) => {
                 console.log("Resposta da API:", response);
-                window.location.reload();
             })
             .catch((error) => {
                 console.log("Erro ao atualizar o usuário:", error);
-            });
+            }).finally(()=> window.location.reload())
     }
     return (
         <>
@@ -91,7 +92,7 @@ export function ContentStartup() {
 
                 <form className="row" onSubmit={handleSubmit(onSubmit)}>
                     <div className="col-lg-4 col-md-12">
-                        <ProfileCArdImageStartup register={register} enterprise={enterpriseDetail} />
+                        <ProfileCArdImageStartup enterprise_id={enterprise_id} token={token} enterprise={enterpriseDetail} />
                         <CardDetails  register={register}/>
                         <PartenrsCard enterprise_id={enterprise_id} token={token}/>
                     </div>
